Handle UMDH "Alias" placeholder token
Sometimes a stack contains a token called "Alias", which I believe UMDH inserts into the stack trace for symbols that can't be resolved for one reason or another. This caused the UMDH Parser to crash with NullReferenceException on Module and other properties that couldn't be set as a result of Alias.

Added bogus Module entry for "Alias" so that downstream code can deal without jumping through hoops.

This fix handles that token g